POC: add discovery of PostgreSQL databases and push them as entities (#7)
* POC: add discovery of PostgreSQL databases and push them as entities

We take all services with label application=spilo and try to connect to the
database cluster using .svc.cluster.local DNS name.

We assume that all Spilo clusters have the same user and password configured
with wich we can connect and list the database names.  The username and the
password need to be provided in two new environment variables, namely:
ZMON_AGENT_POSTGRES_{USER,PASS}.  We assume that we can reuse pre-existing
user that is configured for ZMON worker already.

For every non-system database we create a separate entity of type
'postgresql_database' and configure the 'shards' property so that ZMON checks
can use `sql()` function without further parameters to access the database.

* Check if credentials are specified for PostgreSQL database discovery

Log a warning on initialization if not.

* Add missing build dependency to docker: libpq-dev

Needed to install psycopg2 python package.

* Fix listing postgres databases on connection problems

Also set sslmode to 'require' to avoid 2nd unencrypted connection attempt.

* Remove postgresdb- prefix from PostgreSQL database entity id

Filtering is usually done on the entity type, so having this in the id is not
very helpful.